Modificado application-test.properties e insertado etiquetas en los tests de los repos para que coja dicha configuracion. Tests de los repos tambien hechos y funcionando correctamente.

+++ src/main/java/com/example/fishingapp/dto/FishCaptureDto.java
@@ -8,7 +8,7 @@ public record FishCaptureDto (
      Long id,
      Long userId,
      String fishType,
-     Double weight,
+     Float weight,
      LocalDate captureData,
      String location,
      LocalDateTime createdAt){}
